Solution for Bonus #2
Another new decorator for me: the @methodname.setter

Basically, the key thing I had to get in order to make sense of this was to overload my method.

Weird thing was that what I wanted to do intuitively ended up causing recursion. So I could not put the diameter setting within the overloaded function.